move tutorial turn_events to their own Python module
Requested by Morlic.

I'm intentionally not changing any code in this commit,
not even fixing typos.
